Changing file structure slightly and working on fixing bugs (now it is in the test harness).

+++ Monte/FitnessBased.cs
@@ -48,7 +48,7 @@ namespace Monte
 			float? bestScore = null;
 			foreach(AIState child in children)
 			{
-				child.stateScore = model.evaluate(child.stateRep, child.playerIndex);
+				child.stateScore = (float)model.evaluate(child.stateRep, child.playerIndex);
 				if (bestScore == null ||child.stateScore > bestScore) {
 					best = child;
 					bestScore = child.stateScore;
